+++ spring-aop/src/main/java/org/aopalliance/intercept/ConstructorInterceptor.java
@@ -16,6 +16,8 @@
 
 package org.aopalliance.intercept;
 
+import javax.annotation.Nonnull;
+
 /**
  * Intercepts the construction of a new object.
  *
@@ -54,6 +56,7 @@ public interface ConstructorInterceptor extends Interceptor  {
 	 * @throws Throwable if the interceptors or the target object
 	 * throws an exception
 	 */
+	@Nonnull
 	Object construct(ConstructorInvocation invocation) throws Throwable;
 
 }
