+++ spring-aop/src/main/java/org/aopalliance/intercept/Invocation.java
@@ -16,6 +16,8 @@
 
 package org.aopalliance.intercept;
 
+import javax.annotation.Nonnull;
+
 /**
  * This interface represents an invocation in the program.
  *
@@ -32,6 +34,7 @@ public interface Invocation extends Joinpoint {
 	 * array to change the arguments.
 	 * @return the argument of the invocation
 	 */
+	@Nonnull
 	Object[] getArguments();
 
 }
