+++ spring-aop/src/main/java/org/aopalliance/intercept/MethodInterceptor.java
@@ -16,6 +16,9 @@
 
 package org.aopalliance.intercept;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 /**
  * Intercepts calls on an interface on its way to the target. These
  * are nested "on top" of the target.
@@ -52,6 +55,7 @@ public interface MethodInterceptor extends Interceptor {
 	 * @throws Throwable if the interceptors or the target object
 	 * throws an exception
 	 */
-	Object invoke(MethodInvocation invocation) throws Throwable;
+	@Nullable
+	Object invoke(@Nonnull MethodInvocation invocation) throws Throwable;
 
 }
