+++ spring-aop/src/main/java/org/springframework/aop/interceptor/DebugInterceptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,8 @@
 package org.springframework.aop.interceptor;
 
 import org.aopalliance.intercept.MethodInvocation;
+
+import org.springframework.lang.Nullable;
 
 /**
  * AOP Alliance {@code MethodInterceptor} that can be introduced in a chain
@@ -56,6 +58,7 @@ public class DebugInterceptor extends SimpleTraceInterceptor {
 
 
 	@Override
+	@Nullable
 	public Object invoke(MethodInvocation invocation) throws Throwable {
 		synchronized (this) {
 			this.count++;
