+++ spring-aop/src/test/java/org/springframework/aop/interceptor/ExposedInvocationTestBean.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2002-2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.aop.interceptor;
+
+import org.aopalliance.intercept.MethodInvocation;
+
+import org.springframework.beans.testfixture.beans.TestBean;
+
+abstract class ExposedInvocationTestBean extends TestBean {
+
+	@Override
+	public String getName() {
+		MethodInvocation invocation = ExposeInvocationInterceptor.currentInvocation();
+		assertions(invocation);
+		return super.getName();
+	}
+
+	@Override
+	public void absquatulate() {
+		MethodInvocation invocation = ExposeInvocationInterceptor.currentInvocation();
+		assertions(invocation);
+		super.absquatulate();
+	}
+
+	protected abstract void assertions(MethodInvocation invocation);
+}
