+++ spring-context-support/src/main/java/org/springframework/cache/jcache/interceptor/JCacheInterceptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2018 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,6 +27,7 @@ import org.springframework.cache.interceptor.CacheErrorHandler;
 import org.springframework.cache.interceptor.CacheOperationInvoker;
 import org.springframework.cache.interceptor.SimpleCacheErrorHandler;
 import org.springframework.lang.Nullable;
+import org.springframework.util.Assert;
 import org.springframework.util.function.SingletonSupplier;
 
 /**
@@ -78,8 +79,10 @@ public class JCacheInterceptor extends JCacheAspectSupport implements MethodInte
 			}
 		};
 
+		Object target = invocation.getThis();
+		Assert.state(target != null, "Target must not be null");
 		try {
-			return execute(aopAllianceInvoker, invocation.getThis(), method, invocation.getArguments());
+			return execute(aopAllianceInvoker, target, method, invocation.getArguments());
 		}
 		catch (CacheOperationInvoker.ThrowableWrapper th) {
 			throw th.getOriginal();
