+++ spring-context/src/main/java/org/springframework/cache/interceptor/CacheInterceptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2017 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
 
 import org.springframework.lang.Nullable;
+import org.springframework.util.Assert;
 
 /**
  * AOP Alliance MethodInterceptor for declarative cache
@@ -57,8 +58,10 @@ public class CacheInterceptor extends CacheAspectSupport implements MethodInterc
 			}
 		};
 
+		Object target = invocation.getThis();
+		Assert.state(target != null, "Target must not be null");
 		try {
-			return execute(aopAllianceInvoker, invocation.getThis(), method, invocation.getArguments());
+			return execute(aopAllianceInvoker, target, method, invocation.getArguments());
 		}
 		catch (CacheOperationInvoker.ThrowableWrapper th) {
 			throw th.getOriginal();
