+++ spring-context/src/main/java/org/springframework/cache/interceptor/CacheOperationInvoker.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  */
 
 package org.springframework.cache.interceptor;
+
+import org.springframework.lang.Nullable;
 
 /**
  * Abstract the invocation of a cache operation.
@@ -36,6 +38,7 @@ public interface CacheOperationInvoker {
 	 * @return the result of the operation
 	 * @throws ThrowableWrapper if an error occurred while invoking the operation
 	 */
+	@Nullable
 	Object invoke() throws ThrowableWrapper;
