+++ spring-context/src/main/java/org/springframework/jndi/JndiObjectFactoryBean.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2018 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -368,6 +368,7 @@ public class JndiObjectFactoryBean extends JndiObjectLocator
 		}
 
 		@Override
+		@Nullable
 		public Object invoke(MethodInvocation invocation) throws Throwable {
 			Context ctx = (isEligible(invocation.getMethod()) ? this.jndiTemplate.getContext() : null);
 			try {
